requirements.txt
eaf_base_api
bs4
beautifulsoup4
lxml
m3u8
brotli
av